MAJOR UPDATE:- a whole new scene. read description
pulkit's scene was having FPS drop issues, hence it was required to change it. I was unable to detect the exact cause behind FPS drop. So I made a whole new scene with similar content.
Old scene is now renamed to zoldpulkitportfolio.unity And the new scene is named as Pulkitportfolio.unity
Also, there are some file management fixes as well in this commit.